requirements.txt
aiohttp
anytree
appdirs
aria2p
asgiref
attrdict
beautifulsoup4
cfscrape
feedparser
gc-python-utils
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
lk21
lxml
pillow
psutil
psycopg2-binary
pybase64
pyrogram
python-dotenv
python-magic
python-telegram-bot
qbittorrent-api
requests
speedtest-cli
six
telegraph
tenacity
TgCrypto
torrentool==1.1.0
urllib3
yt-dlp
lxml
megasdkrestclient